Made rmrf tolerant of files or dirs that do not exist.

+++ rmrf.ps1
@@ -1,5 +1,11 @@
 #!/usr/bin/env pwsh
 
-$dirOrFile=$args[0]
-Write-Host "Remove-Item -Recurse -Force $dirorFile"
-Remove-Item -Recurse -Force $dirOrFile
+$dirOrFile = $args[0]
+
+if (test-path $dirOrFile) {
+    Write-Host "Remove-Item -Recurse -Force $dirOrFile"
+    Remove-Item -Recurse -Force $dirOrFile
+}
+else {
+    Write-Host "$dirOrFile does not exist."
+}
